requirements.txt
flask
supabase
requests
python-dotenv
Flask-Caching
google-cloud-storage
google-auth